poc/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pydantic
pdfminer.six
pandas
openpyxl
xlrd
openai
python-dotenv
